+++ main.py
@@ -12,6 +12,15 @@ app = FastAPI(max_upload_size=50_000_000)
 
 # Modelo de datos para el request (usando diccionario, luego puedes usar Pydantic)
 
+@app.get("/prueba.php/")
+def pruebas(idOwner: int = Query(..., alias="owner")):
+    connection = get_db_connection()
+    fotos = getParkingByOwner(connection, idOwner)
+    return {"parqueaderos": fotos}
+
+
+
+
 @app.post(PATH_NEW_USER)
 def crearusuario(user: dict):
     connection = get_db_connection()
@@ -106,6 +115,45 @@ def cities(stateId: int = Query(..., example=1, description="id del paiestados",
     finally:
         connection.close()
 
+@app.get(PATH_GET_PARKING)
+def getParking(idParking: int = Query(..., alias="parqueadero")):
+    connection = get_db_connection()
+    if not connection:
+        raise HTTPException(status_code=500, detail=DB_CONECCTION_ERROR)
+    try:
+        parkings = getParkingByID(connection, idParking)
+        if not parkings:
+            raise HTTPException(
+                status_code=404,
+                detail=f"No se encontraron Datos para el parqueadero {idParking}"
+            )
+        return parkings
+    except pymysql.Error as e:
+        connection.rollback()
+        raise HTTPException(status_code=400, detail=f"Error : {e} revisar")
+    finally:
+        connection.close()
+
+@app.get(PATH_GET_PARKING_OWNER)
+def getPArkings(idOwner: int = Query(..., alias="owner")):
+    connection = get_db_connection()
+    if not connection:
+        raise HTTPException(status_code=500, detail=DB_CONECCTION_ERROR)
+    try:
+        parkings = getParkingByOwner(connection, idOwner)
+        if not parkings:
+            raise HTTPException(
+                status_code=404,
+                detail=f"El usuario {idOwner} no posee parqueaderos"
+            )
+        return {"parqueaderos": parkings}
+    except pymysql.Error as e:
+        connection.rollback()
+        raise HTTPException(status_code=400, detail=f"Error : {e} revisar")
+    finally:
+        connection.close()    
+
+
 @app.post(PATH_NEW_PARKING)
 async def createParking(owner: int = Form(...), place: int = Form(...), address: str = Form(...),
     long: float = Form(...), width: float = Form(...), price: float = Form(...),
@@ -115,6 +163,11 @@ async def createParking(owner: int = Form(...), place: int = Form(...), address:
     if not connection:
         raise HTTPException(status_code=500, detail=DB_CONECCTION_ERROR)
     try:
+        if not files:
+            raise HTTPException(
+                status_code=400,
+                detail="Debe subir al menos una foto del parqueadero"
+            )
         parking_data = {
             OWNER: owner,
             PLACE: place,